% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfino-package.R
\docType{package}
\name{kfino}
\alias{kfino}
\alias{kfino-package}
\title{kfino: Kalman Filtering}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A method for detecting outliers with a Kalman filter on impulsed noised outliers and prediction on cleaned data. 'kfino' is a robust sequential algorithm allowing to filter data with a large number of outliers. This algorithm is based on simple latent linear Gaussian processes as in the Kalman Filter method and is devoted to detect impulse-noised outliers. These are data points that differ significantly from other observations. 'ML' (Maximization Likelihood) and 'EM' (Expectation-Maximization algorithm) algorithms were implemented in 'kfino'. The method is described in full details in the following arXiv e-Print: \href{https://arxiv.org/abs/2208.00961}{arXiv:2208.00961}.
}
\details{
xxxxxxxx xxxxxxxxxxxxxxx xxxxxxxxxxxxxxxxx xxxxxxxxxxxxxxxxx.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://forgemia.inra.fr/isabelle.sanchez/kfino}
  \item Report bugs at \url{https://forgemia.inra.fr/isabelle.sanchez/kfino/-/issues}
}

}
\author{
\strong{Maintainer}: Isabelle Sanchez \email{isabelle.sanchez@inrae.fr}

Authors:
\itemize{
  \item Bertrand Cloez \email{bertrand.cloez@inrae.fr}
}

Other contributors:
\itemize{
  \item Benedicte Fontez \email{benedicte.fontez@supagro.fr} [contractor]
}

}
