% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merinos2}
\alias{merinos2}
\title{a dataset containing the WoW weighing for one animal (merinos lamb) of 345
observations, difficult to model. https://doi.org/10.1016/j.compag.2018.08.022}
\format{
a data.frame
\describe{
  \item{Poids}{weight (in kg)}
  \item{Date}{Date of weighing yyyy-mm-dd}
  \item{IDE}{id of the animal}
  \item{Day}{Date of weighing with day and time yyyy-mm-dd hh:mm:ss}
  \item{dateNum}{a rescaled date - fraction of the whole observational time 
  for one individual. 
  \eqn{dateNum=(Heure - min(Heure))/86400 + (Date - min(Date))/86400}}
}
}
\usage{
merinos2
}
\description{
A dataset for kfino algorithm
}
\keyword{datasets}
